+++ app/forms/transfer/task/deed_of_novation_and_variation_task_form.rb
@@ -1,0 +1,9 @@
+class Transfer::Task::DeedOfNovationAndVariationTaskForm < BaseOptionalTaskForm
+  attribute :received, :boolean
+  attribute :cleared, :boolean
+  attribute :signed_outgoing_trust, :boolean
+  attribute :signed_incoming_trust, :boolean
+  attribute :saved, :boolean
+  attribute :signed_secretary_state, :boolean
+  attribute :sent, :boolean
+end