+++ app/models/transfer/task_list.rb
@@ -11,7 +11,8 @@ class Transfer::TaskList < ::BaseTaskList
       {
         identifier: :legal_documents,
         tasks: [
-          Transfer::Task::MasterFundingAgreementTaskForm
+          Transfer::Task::MasterFundingAgreementTaskForm,
+          Transfer::Task::DeedOfNovationAndVariationTaskForm
         ]
       }
     ]